better solution
